% Generated by roxygen2: do not edit by hand
% Please edit documentation in featurization/countpattern.R
\name{countpattern}
\alias{countpattern}
\title{Illustration of countpattern
This function takes sequence and pattern as input and count how many times a particular pattern is present in the sequence.}
\usage{
countpattern(sequence, pattern)
}
\arguments{
\item{sequence}{provided as a list of sequences}

\item{pattern}{a string}
}
\value{
a list of integer indicating frequency of pattern.
}
\description{
Illustration of countpattern
This function takes sequence and pattern as input and count how many times a particular pattern is present in the sequence.
}
\examples{
sequence = list("ABDEFGHABDAACBBDEBGGGHHH", "ABCBDBEBEBBBDBDBFDFDFGGHHEEFFEECCCD")
pattern = "BD"
feat = featurization(sequence, pattern)
feat
}

