% Generated by roxygen2: do not edit by hand
% Please edit documentation in crisprpred_main/crisprpred_main.R
\name{crisprpred_main}
\alias{crisprpred_main}
\title{Explanation of crisprpred_main functions}
\usage{
crisprpred_main(datasetpath, featurelist, kfold, iteration4dl, trees,
  learningrate, samplingrate)
}
\arguments{
\item{datasetpath}{full path of a csv file}

\item{featurelist}{provided by user as a list of strings}

\item{kfold}{used in ML-functions for kfold cross-validation}

\item{iteration4dl}{number of time dataset will be iterated}

\item{trees}{number of trees in random forest}

\item{learningrate}{learning rate of deep learner}

\item{samplingrate}{sampling rate in random forest}
}
\value{
None
}
\description{
This function takes full datasetpath and reads data as a R object (data frame), a list of features and a number to denote cross-validation. It also takes other parameters for different algorithms. Then, it performs Machine Learning algorithms and build prediction models. Then it predicts sgRNA activity based on prediction models.
}
\examples{
setwd('..')
#suppose we have a file as '../crisprpred/data-raw/sample.csv' and current directory is set to '../crisprpred'
dir = getwd()
datasetpath = paste0(dir,'/data-raw/sample.csv')
featurelist = c("X30mer", "Percent.Peptide", "Amino.Acid.Cut.position","predictions")
kfoldCross = 2
crisprpred_main(datasetpath, featurelist, kfoldCross, 3, 4, 0.66)
}

