% Generated by roxygen2: do not edit by hand
% Please edit documentation in mlmethods/dplearning.R
\name{dplearning}
\alias{dplearning}
\title{Deep Learning}
\usage{
dplearning(featurelist, featuredata, leaveonegene = 0, kfold = 10,
  learningrate = 0.6)
}
\arguments{
\item{featurelist}{a list of features. last name will indicate the value to be predicted.}

\item{featuredata}{a sample dataset containing all features}

\item{leaveonegene}{check for leaveonegeneout cross-validation}

\item{kfold}{a value for cross validation. Default value is 3.}

\item{learningrate}{a fractional learning rate for deep learner. Default value is 0.6.}
}
\value{
spearman correlation
}
\description{
This function takes full filepath, a list of learning features, a value for cross-validation, the number of times data set will be iterated and learning rate. Now, it creates a deep learning model using  deeplearning function of h2o package and outputs RMSE based on provided dataset. Note that size of dataset should be enough to choose a suitable value for kfold.
}
\examples{
featurelist = c("Percent.Peptide", "Amino.Acid.Cut.position","predictions")
#suppose we have a file as '../crisprpred/data-raw/sample.csv' and current directory is set to '../crisprpred'
dir = getwd()
filepath = paste0(dir,'/data-raw/sample.csv')
data = read.csv(filepath)
dplearning(featurelist,data)
}

