% Generated by roxygen2: do not edit by hand
% Please edit documentation in featureformula/featureformula.R
\name{featureformula}
\alias{featureformula}
\title{Making Formula for Learning}
\usage{
featureformula(featurelist)
}
\arguments{
\item{featurelist}{a list of strings}
}
\value{
a formula
}
\description{
This function takes a list of features to make a suitable formula. For example, it creates a formula Y~X1+X2+X3 from a list of features ('X1','X2','X3','Y')
}
\examples{
featurelist = c('X1', 'X2', 'X3', 'Y')
formula = featureformula(featurelist)
formula
}

