% Generated by roxygen2: do not edit by hand
% Please edit documentation in featurization/featurization.R
\name{featurization}
\alias{featurization}
\title{Illustration of Featurization}
\usage{
featurization(sequences, string, seq = TRUE, seqorder = 2, pos = TRUE,
  posorder = 2)
}
\arguments{
\item{sequences}{provided as dataframe}

\item{string}{a list of aminoacids or nucleotides}

\item{seq}{sequence based features. by default it is true.}

\item{seqorder}{highest number of sequence which will be considered together}

\item{pos}{position specific features. by default it is true.}

\item{posorder}{highest number of sequence which will be considered together}
}
\value{
a featurized dataframe
}
\description{
This function takes dataset and a list of features as input and produce a features-wise dataset. The number of columns in returned dataset is equal to the number of features in featurelist.
}
\examples{
input = list("ABCDEFGHABDAACBBDEBGGGHHH", "ABCBDBEBEBBBDBDBFDFDFGGHHEEFFEECCCD")
string = c("A", "BD")
featuredata = featurization(input, string, seq = TRUE, pos = FALSE)
featuredata
}

