% Generated by roxygen2: do not edit by hand
% Please edit documentation in featurization/findposition.R
\name{findposition}
\alias{findposition}
\title{Illustration of findposition
This function takes sequence, pattern and position as input and check whether a particulaer pattern is present in position-th place of sequence.}
\usage{
findposition(sequence, pattern, position)
}
\arguments{
\item{sequence}{provided as a list of sequences}

\item{pattern}{a string}

\item{position}{an integer value}
}
\value{
a list of 0/1 indicating present or absent.
}
\description{
Illustration of findposition
This function takes sequence, pattern and position as input and check whether a particulaer pattern is present in position-th place of sequence.
}
\examples{
sequence = list("ABDEFGHABDAACBBDEBGGGHHH", "ABCBDBEBEBBBDBDBFDFDFGGHHEEFFEECCCD")
pattern = "BD"
position = 2
feat = featurization(sequence, pattern, position)
feat
}

