% Generated by roxygen2: do not edit by hand
% Please edit documentation in mlmethods/lmregression.R
\name{lmregression}
\alias{lmregression}
\title{Linear Regression}
\usage{
lmregression(featurelist, featuredata, leaveonegene = 0, kfold = 10)
}
\arguments{
\item{featurelist}{a list of feature}

\item{featuredata}{provided dataset}

\item{leaveonegene}{check for leaveonegeneout cross-validation}

\item{kfold}{a value for cross validation}
}
\value{
spearman correlation
}
\description{
This function takes featurelist, dataset and a value for cross-validation. Now, it creates a formula and outputs RMSE based on provided dataset.
}
\examples{
featurelist = c("Percent.Peptide", "Amino.Acid.Cut.position","predictions")
dir = getwd()
filepath = paste0(dir,'/data-raw/sample.csv')
data = read.csv(filepath)
lmregression(featurelist,data,0)
}

