% Generated by roxygen2: do not edit by hand
% Please edit documentation in mlmethods/mars.R
\name{mars}
\alias{mars}
\title{MARS Regression}
\usage{
mars(featurelist, featuredata, leaveonegene = 0, kfold = 10)
}
\arguments{
\item{featurelist}{a list of features}

\item{featuredata}{provided dataset}

\item{leaveonegene}{check for leaveonegeneout cross-validation}

\item{kfold}{a value for cross validation}
}
\value{
spearman correlation
}
\description{
This function takes mars regression formula, dataset and a value for cross-validation. Now, it outputs RMSE based on provided dataset.
}
\examples{
featurelist = c("Percent.Peptide", "Amino.Acid.Cut.position","predictions")
dir = getwd()
filepath = paste0(dir,'/data-raw/sample.csv')
data = read.csv(filepath)
mars(featurelist,data,0,2)
}

