% Generated by roxygen2: do not edit by hand
% Please edit documentation in mlmethods/randomforest0.R
\name{randomforest0}
\alias{randomforest0}
\title{Random Forest}
\usage{
randomforest0(featurelist, featuredata, leaveonegene = 0, kfold = 10,
  trees = 500)
}
\arguments{
\item{featurelist}{a list of features. last name will indicate the value to be predicted.}

\item{featuredata}{a sample dataset containing all features}

\item{leaveonegene}{check for leaveonegeneout cross-validation}

\item{kfold}{a value for cross validation. Default value is 3.}

\item{trees}{number of trees that will be built.}
}
\value{
spearman correlation
}
\description{
This function takes full filepath, a list of learning features, a value for cross-validation, the number of times data set will be iterated and learning rate. Now, it creates a deep learning model using  deeplearning function of h2o package and outputs RMSE based on provided dataset. Note that size of dataset should be enough to choose a suitable value for kfold.
}
\examples{
featurelist = c("Percent.Peptide", "Amino.Acid.Cut.position","predictions")
#suppose we have a file as '../crisprpred/data-raw/sample.csv' and current directory is set to '../crisprpred'
dir = getwd()
filepath = paste0(dir,'/data-raw/sample.csv')
data = read.csv(filepath)
randomforest(featurelist,data,leaveonegene = 0)
}

