% Generated by roxygen2: do not edit by hand
% Please edit documentation in metrics/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(error)
}
\arguments{
\item{error}{a value denoting error}
}
\value{
rmse-error
}
\description{
Return square root of mean squared error.
}
\examples{
rmse(5)
}

