% Generated by roxygen2: do not edit by hand
% Please edit documentation in featurization/viennaRNADataManipulation.R
\name{viennaRNADataManipulation}
\alias{viennaRNADataManipulation}
\title{Description of viennaRNADataManipulation Function}
\usage{
viennaRNADataManipulation(sequences)
}
\arguments{
\item{sequences}{a list of sequence strings}
}
\value{
datafram of extracted features based on input sequences
}
\description{
This function takes a list of sequences as input, manipulates data using rna sequence of viennaRNA and returns a dataframe.
}
\examples{
s = c('AGGCGTGTTAACT','ACGTTTAAGCT')
viennaRNADataManipulation(s)
}

